+++ src/pages/Groups.tsx
@@ -10,6 +10,7 @@ import { Card, CardContent, CardFooter, CardHeader } from "@/components/ui/card"
 import { GroupCard } from "@/components/groups/GroupCard";
 import { GroupSearch } from "@/components/groups/GroupSearch";
 import { useState } from "react";
+import { useGroupStats } from "@/hooks/useGroupStats";
 
 export default function Groups() {
   const { nostr } = useNostr();
@@ -28,46 +29,7 @@ export default function Groups() {
   });
 
   // Query for community stats (posts and participants)
-  const { data: communityStats, isLoading: isLoadingStats } = useQuery({
-    queryKey: ["community-stats"],
-    queryFn: async (c) => {
-      if (!communities || communities.length === 0) return {};
-
-      const signal = AbortSignal.any([c.signal, AbortSignal.timeout(8000)]);
-      const stats: Record<string, { posts: number; participants: Set<string> }> = {};
-
-      // Create a filter for all communities to get posts in a single query
-      const communityRefs = communities.map(community => {
-        const dTag = community.tags.find(tag => tag[0] === "d");
-        return `34550:${community.pubkey}:${dTag ? dTag[1] : ""}`;
-      });
-
-      // Get all posts that reference any community
-      const posts = await nostr.query([{
-        kinds: [1, 4550],
-        "#a": communityRefs,
-        limit: 500
-      }], { signal });
-
-      // Process posts to get stats for each community
-      for (const post of posts) {
-        const communityTag = post.tags.find(tag => tag[0] === "a");
-        if (!communityTag) continue;
-
-        const communityId = communityTag[1];
-        if (!stats[communityId]) {
-          stats[communityId] = { posts: 0, participants: new Set() };
-        }
-
-        // Count posts and unique participants
-        stats[communityId].posts++;
-        stats[communityId].participants.add(post.pubkey);
-      }
-
-      return stats;
-    },
-    enabled: !!nostr && !!communities && communities.length > 0,
-  });
+  const { data: communityStats, isLoading: isLoadingStats } = useGroupStats(communities);
 
   // Filter communities based on search query
   const filteredCommunities = communities?.filter(community => {
